Refactor: Extract FromSql into its own module.
